+++ src/main/java/fr/gino/lsftranslate/repository/CatGramRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CatGramRepository extends JpaRepository<CatGram, Long> {
+public interface CatGramRepository extends JpaRepository<CatGram, Integer> {
 }
