+++ src/main/java/fr/gino/lsftranslate/repository/LettreRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LettreRepository extends JpaRepository<Lettre, Long> {
+public interface LettreRepository extends JpaRepository<Lettre, Integer> {
 }
