+++ src/main/java/fr/gino/lsftranslate/service/LettreService.java
@@ -5,5 +5,5 @@ import fr.gino.lsftranslate.model.Lettre;
 import java.util.Optional;
 
 public interface LettreService {
-    Optional<Lettre> findById(long id);
+    Optional<Lettre> findById(int id);
 }
