+++ src/main/java/fr/gino/lsftranslate/service/MotService.java
@@ -5,5 +5,5 @@ import fr.gino.lsftranslate.model.Mot;
 import java.util.Optional;
 
 public interface MotService {
-    Optional<Mot> findById(long id);
+    Optional<Mot> findById(int id);
 }
